requirements.txt
Django
djangorestframework
Pillow
djangorestframework-simplejwt
psycopg2-binary